Time: 49 ms (80.52%) | Memory: 125 MB (43.53%) - LeetSync

+++ 295-find-median-from-data-stream/find-median-from-data-stream.cpp
@@ -27,7 +27,7 @@ public:
     double findMedian() {
         double median = 0;
         if (leftHeap.size() == rightHeap.size()) {
-            median = (double) (leftHeap.top() + rightHeap.top()) / 2;
+            median = (leftHeap.top() + rightHeap.top()) / 2.0;
         } else {
             median = leftHeap.top();
         }
